refactor(lib): group utils into args/, selection/, chrome/, daemon/
lib/utils had grown into a 37-file junk drawer. Peel the cohesive
clusters out into domain folders alongside the existing feature dirs:

- lib/args/            CLI argument parsing (tokenize-args, parse-cli-flags)
- lib/selection/      test matching for -t / #line / --search
                      (parse-test-declarations, line-targets, qunit-filter,
                      qunit-filter-match)
- lib/chrome/         Chromium launch machinery (prelaunch, spawn, args,
                      find, cleanup-dir) — the chrome-/chromium- prefix is
                      dropped now that the folder implies it
- lib/commands/daemon/  daemon-only helpers (parse-idle-timeout, hint,
                      socket-path) moved beside the daemon command itself

Cross-cutting primitives stay in utils (source-map-decoder is used by the
reporter and coverage; kill-process-group and open-output-in-browser are
generic / browser-general). Pure git mv + import-path rewrites; no exported
symbols renamed. Test files moved to mirror the new layout.
