+++ lib/chrome/cleanup-dir.ts
