+++ lib/commands/daemon/hint.ts
