+++ lib/commands/daemon/parse-idle-timeout.ts
