+++ lib/commands/daemon/socket-path.ts
