+++ lib/selection/line-targets.ts
