+++ lib/setup/config.ts
@@ -8,7 +8,7 @@ import { findProjectRoot } from '../utils/find-project-root.ts';
 import { buildFSTree } from './fs-tree.ts';
 import { setupTestFilePaths } from './test-file-paths.ts';
 import { getChangedFsTree } from './get-changed-fs-tree.ts';
-import { parseCliFlags } from '../utils/parse-cli-flags.ts';
+import { parseCliFlags } from '../args/parse-cli-flags.ts';
 import { resolveOnlyFailedFiles, type FailedTestRecord } from '../utils/failure-cache.ts';
 import { createReporters } from '../reporter/index.ts';
 import type { Config, FSTree } from '../types.ts';
