+++ lib/utils/open-output-in-browser.ts
@@ -1,8 +1,8 @@
 import { spawn } from 'node:child_process';
 import path from 'node:path';
 import { pathToFileURL } from 'node:url';
-import { findChrome } from './find-chrome.ts';
-import { qunitFilterQuery } from './qunit-filter.ts';
+import { findChrome } from '../chrome/find.ts';
+import { qunitFilterQuery } from '../selection/qunit-filter.ts';
 import type { Config } from '../types.ts';
 
 /**
