+++ test/commands/daemon/socket-path-test.ts
@@ -1,8 +1,12 @@
 import { module, test } from 'qunitx';
 import os from 'node:os';
 import path from 'node:path';
-import { daemonSocketPath, daemonInfoPath, daemonDir } from '../../lib/utils/daemon-socket-path.ts';
-import '../helpers/custom-asserts.ts';
+import {
+  daemonSocketPath,
+  daemonInfoPath,
+  daemonDir,
+} from '../../../lib/commands/daemon/socket-path.ts';
+import '../../helpers/custom-asserts.ts';
 
 const FIXED_CWD = '/some/test/project';
